backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-celery-results==2.4.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
django-extensions==3.2.3

# AI/ML Libraries
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
ultralytics==8.0.200
opencv-python==4.8.1.78
librosa==0.10.1
tensorflow==2.16.1
scikit-learn==1.3.0
xgboost==1.7.6
prophet==1.1.4
transformers==4.35.2
sentencepiece==0.2.0
tokenizers==0.15.0
arabert==1.0.1
farasapy==0.0.7
camel-tools>=1.5.4

pyarabic==0.6.15
SpeechRecognition==3.10.0

# Audio Processing
pydub==0.25.1
audioread==3.0.1
webrtcvad==2.0.10
soundfile==0.12.1

# Database & API
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
Pillow==10.0.1
scipy==1.11.3
plotly==5.17.0

# Real-time Processing
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
websockets==12.0

# Security & Authentication
cryptography==41.0.7
pyjwt==2.8.0
bcrypt==4.0.1

# Utilities
python-multipart==0.0.6
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
tqdm==4.66.1

# Geo Spatial
geopy==2.3.0
folium==0.14.0

# Monitoring
prometheus-client==0.17.1
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.13.2
